+++ tests/test_avg.c
@@ -10,7 +10,7 @@ static void test_mm_avg_epu8(void **state)
 {
 	(void) state;
 
-	__m128i a = _mm_setr_epi8(0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 110, 120, 130, 140, 150);
+	__m128i a = _mm_setr_epi8(0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 110, 120, (int8_t) 130, (int8_t) 140, (int8_t) 150);
 	__m128i b = _mm_setr_epi8(0, 4, 8, 12, 16, 20, 24, 28, 32, 36, 40, 44, 48, 52, 56, 60);
 
 	__m128i result = _mm_avg_epu8(a, b);
