+++ tests/test_cvtep.c
@@ -131,7 +131,7 @@ static void test_mm_cvtepu8_epi16(void **state)
 	(void) state;
 
 #ifdef XSSE4_1
-	__m128i x = _mm_setr_epi8(10, 20, 30, 40, 50, 60, 70, UINT8_MAX, 11, 12, 13, 14, 15, 16, 17, 18);
+	__m128i x = _mm_setr_epi8(10, 20, 30, 40, 50, 60, 70, (int8_t) UINT8_MAX, 11, 12, 13, 14, 15, 16, 17, 18);
 	__m128i result = _mm_cvtepu8_epi16(x);
 
 	int16_t expected[8] = { 10, 20, 30, 40, 50, 60, 70, UINT8_MAX };
@@ -151,7 +151,7 @@ static void test_mm_cvtepu8_epi32(void **state)
 	(void) state;
 
 #ifdef XSSE4_1
-	__m128i x = _mm_setr_epi8(10, 20, 30, UINT8_MAX, 50, 60, 70, 80, 11, 12, 13, 14, 15, 16, 17, 18);
+	__m128i x = _mm_setr_epi8(10, 20, 30, (int8_t) UINT8_MAX, 50, 60, 70, 80, 11, 12, 13, 14, 15, 16, 17, 18);
 	__m128i result = _mm_cvtepu8_epi32(x);
 
 	int32_t expected[4] = { 10, 20, 30, UINT8_MAX };
@@ -171,7 +171,7 @@ static void test_mm_cvtepu8_epi64(void **state)
 	(void) state;
 
 #ifdef XSSE4_1
-	__m128i x = _mm_setr_epi8(10, UINT8_MAX, 30, 40, 50, 60, 70, 80, 11, 12, 13, 14, 15, 16, 17, 18);
+	__m128i x = _mm_setr_epi8(10, (int8_t) UINT8_MAX, 30, 40, 50, 60, 70, 80, 11, 12, 13, 14, 15, 16, 17, 18);
 	__m128i result = _mm_cvtepu8_epi64(x);
 
 	int64_t expected[2] = { 10, UINT8_MAX };
@@ -191,7 +191,7 @@ static void test_mm_cvtepu16_epi32(void **state)
 	(void) state;
 
 #ifdef XSSE4_1
-	__m128i x = _mm_setr_epi16(10, 20, 30, UINT16_MAX, 50, 60, 70, 80);
+	__m128i x = _mm_setr_epi16(10, 20, 30, (int16_t) UINT16_MAX, 50, 60, 70, 80);
 	__m128i result = _mm_cvtepu16_epi32(x);
 
 	int32_t expected[4] = { 10, 20, 30, UINT16_MAX };
@@ -211,7 +211,7 @@ static void test_mm_cvtepu16_epi64(void **state)
 	(void) state;
 
 #ifdef XSSE4_1
-	__m128i x = _mm_setr_epi16(10, UINT16_MAX, 30, 40, 50, 60, 70, 80);
+	__m128i x = _mm_setr_epi16(10, (int16_t) UINT16_MAX, 30, 40, 50, 60, 70, 80);
 	__m128i result = _mm_cvtepu16_epi64(x);
 
 	int64_t expected[2] = { 10, UINT16_MAX };
@@ -231,7 +231,7 @@ static void test_mm_cvtepu32_epi64(void **state)
 	(void) state;
 
 #ifdef XSSE4_1
-	__m128i x = _mm_setr_epi32(10, UINT32_MAX, 30, 40);
+	__m128i x = _mm_setr_epi32(10, (int32_t) UINT32_MAX, 30, 40);
 	__m128i result = _mm_cvtepu32_epi64(x);
 
 	int64_t expected[2] = { 10, UINT32_MAX };
