+++ tests/test_mul.c
@@ -83,7 +83,7 @@ static void test_mm_maddubs_epi16(void **state)
 	(void) state;
 
 #ifdef XSSSE3
-	__m128i a = _mm_setr_epi8(10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 110, 120, 130, 140, 150, 160);
+	__m128i a = _mm_setr_epi8(10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 110, 120, (int8_t) 130, (int8_t) 140, (int8_t) 150, (int8_t) 160);
 	__m128i b = _mm_setr_epi8(1, -2, 3, -4, 5, -6, 7, -8, 9, -10, 11, -12, 13, 14, 15, 16);
 
 	__m128i result = _mm_maddubs_epi16(a, b);
